requirements.txt
asn1crypto==0.24.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap-form==3.4
django-heroku==0.3.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
google-auth==1.6.3
gunicorn==19.9.0
idna==2.8
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2